requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages

apache-airflow-providers-snowflake==5.3.1
astronomer-cosmos[dbt-snowflake]==1.3.2
